+++ react-components/src/components/hooks/useSuggestedMarkets.js
@@ -1,19 +1,34 @@
 import { useState, useEffect } from 'react'
 import Services from '@src/Services'
 import { useUserProducts } from '@src/components/hooks/useUserData'
+import { get } from '@src/Helpers'
 
 // Returns the list of suggested markets based on the list of products provided.
 // If no product list provided - use basket instead
+
+const capitalize = (str) =>
+  `${str.charAt(0).toUpperCase()}${str.slice(1).toLowerCase()}`
+
+let localState = {}
 
 export const useSuggestedMarkets = (products) => {
   const {
     products: userProducts,
     loadProducts: loadUserProducts,
+    productsLoaded: userProductsLoaded,
   } = useUserProducts(false)
-  const [suggestedCountries, setSuggestedCountries] = useState({})
+  const [suggestedCountries, setSuggestedCountries] = useState(
+    'suggested_countries'
+  )
+
+  const setLocalState = (newState) => {
+    localState = { ...localState, ...newState }
+    if (localState.details && localState.hs2Desc)
+      setSuggestedCountries(localState)
+  }
 
   const getHS2Code = () => {
-    if (!products && !userProducts.length) {
+    if (!products && !userProductsLoaded) {
       loadUserProducts()
     }
     const productList = products || userProducts || []
@@ -25,6 +40,7 @@ export const useSuggestedMarkets = (products) => {
         (scanProduct) => (scanProduct.commodity_code || '').substr(0, 2) !== hs2
       )
     return {
+      hs6: product.commodity_code,
       hs2,
       product: product.commodity_name,
       allSame,
@@ -33,21 +49,32 @@ export const useSuggestedMarkets = (products) => {
 
   const loadSuggestedCountries = () => {
     const activeHs2 = getHS2Code(products)
+
     if (activeHs2) {
-      if ((suggestedCountries || {}).hs2 !== activeHs2.hs2) {
+      if (localState.hs2 !== activeHs2.hs2) {
+        setLocalState({
+          hs2: activeHs2.hs2,
+          details: activeHs2,
+        })
         Services.getSuggestedCountries(activeHs2.hs2).then((result) => {
-          setSuggestedCountries({
-            hs2: activeHs2.hs2,
-            details: activeHs2,
+          setLocalState({
             suggestions: result,
           })
         })
-      } else {
-        setSuggestedCountries({
-          ...suggestedCountries,
-          details: activeHs2,
-        })
+        Services.lookupProductSchedule({ hsCode: activeHs2.hs6 }).then(
+          (results) => {
+            setLocalState({
+              hs2Desc: capitalize(
+                get(results, 'children.0.children.0.desc', '').replace(
+                  /^chapter[\d\s-]*/i,
+                  ''
+                ) || ''
+              ),
+            })
+          }
+        )
       }
+      setSuggestedCountries(localState)
     }
   }